services/srv-usuarios/requirements.txt
fastapi
uvicorn[standard]

# Dependências para banco de dados e ORM
sqlalchemy
psycopg2-binary

# Dependências para segurança de senha
passlib[bcrypt]

# Dependência opcional do Pydantic para validação de e-mail
pydantic[email]

# Dependências para autenticação JWT e OAuth2
python-jose[cryptography]
python-multipart

# Dependência para a ferramenta de linha de comando
typer[all]

pydantic-settings